requirements.txt
fastapi
uvicorn
starlette
python-multipart
websockets
aiofiles
requests

faster-whisper
ctranslate2
openai-whisper
soundfile
librosa
numpy
scipy

torch
torchvision
torchaudio
transformers
accelerate
bitsandbytes
sentencepiece
safetensors

tts==0.22.0
edge-tts
ffmpeg-python

pandas
scikit-learn
tenacity
pydantic
tqdm
typing-extensions

rich
mutagen>=1.47
librosa>=0.10
resampy>=0.4.3
python-vlc>=3.0.0
